+++ core/ContourManager.py
