+++ core/ImageManager.py
